extract conflict handling into trait
