+++ app/Filament/Server/Pages/ServerFormPage.php
@@ -3,6 +3,7 @@
 namespace App\Filament\Server\Pages;
 
 use App\Models\Server;
+use App\Traits\Filament\BlockAccessInConflict;
 use Filament\Facades\Filament;
 use Filament\Forms\Concerns\InteractsWithForms;
 use Filament\Forms\Form;
@@ -14,6 +15,7 @@ use Filament\Pages\Page;
  */
 abstract class ServerFormPage extends Page
 {
+    use BlockAccessInConflict;
     use InteractsWithFormActions;
     use InteractsWithForms;
 
@@ -63,18 +65,5 @@ abstract class ServerFormPage extends Page
         $server = Filament::getTenant();
 
         return $server;
-    }
-
-    // TODO: find better way handle server conflict state
-    public static function canAccess(): bool
-    {
-        /** @var Server $server */
-        $server = Filament::getTenant();
-
-        if ($server->isInConflictState()) {
-            return false;
-        }
-
-        return parent::canAccess();
     }
 }
