+++ app/Filament/Server/Resources/BackupResource.php
@@ -6,6 +6,7 @@ use App\Filament\Server\Resources\BackupResource\Pages;
 use App\Models\Backup;
 use App\Models\Permission;
 use App\Models\Server;
+use App\Traits\Filament\BlockAccessInConflict;
 use App\Traits\Filament\CanCustomizePages;
 use App\Traits\Filament\CanCustomizeRelations;
 use Filament\Facades\Filament;
@@ -14,6 +15,7 @@ use Illuminate\Database\Eloquent\Model;
 
 class BackupResource extends Resource
 {
+    use BlockAccessInConflict;
     use CanCustomizePages;
     use CanCustomizeRelations;
 
@@ -51,19 +53,6 @@ class BackupResource extends Resource
 
         return $count >= $limit ? 'danger'
             : ($count >= $limit * self::WARNING_THRESHOLD ? 'warning' : 'success');
-    }
-
-    // TODO: find better way handle server conflict state
-    public static function canAccess(): bool
-    {
-        /** @var Server $server */
-        $server = Filament::getTenant();
-
-        if ($server->isInConflictState()) {
-            return false;
-        }
-
-        return parent::canAccess();
     }
 
     public static function canViewAny(): bool
